Restructured slightly
