+++ model/src/form/utils/index.ts
@@ -1,2 +1,3 @@
 export { findStartPage } from '~/src/form/utils/find-start-page.js'
 export { findDefinitionListFromComponent } from '~/src/form/utils/list.js'
+export { preventUnicodeInEmail } from '~/src/form/utils/prevent-unicode.js'
